Conflicts:
	CRMProject/src/features/consoleView/AddContact.java
	CRMProject/src/features/consoleView/ContactsMenu.java
	CRMProject/src/features/consoleView/DisplayContacts.java
	CRMProject/src/features/consoleView/FeatureAppMain.java
	CRMProject/src/features/model/AllContacts.java
	CRMProject/src/features/model/Contact.java

+++ CRMProject/src/features/model/Contact.java
@@ -12,10 +12,10 @@ package features.model;
 public class Contact {
     
     private final String name;
-    private final int phone;
+    private final long phone;
     private final String email;
     
-      public Contact(String _name, int _phone, String _email) {
+      public Contact(String _name, long _phone, String _email) {
         name = _name;
         phone = _phone;
         email = _email;
@@ -23,9 +23,9 @@ public class Contact {
     
     public String getName() {
         return name;
-}
+    }
     
-    public int getPhone() {
+    public long getPhone() {
         return phone;
     }
